popp-up border update
a simple border to the popups was added. Same exact as the ones in regular images. The amount of time it took me to realize how to this do is... kinda silly. Sorta assumed border wouldn't work because of the entire background being blurred and that taking up the whole screen. Well here we are anyway :). AHHHHHHHH